+++ default-assets/branding.js
@@ -1,0 +1,12 @@
+// These default branding values are used when no branding is present in stripes.config.js
+// Default src are prefixed with '@folio/stripes-core' so they can be imported
+// and processed through the webpack loaders from the stripes-core in use.
+module.exports = {
+  logo: {
+    src: '@folio/stripes-core/default-assets/folio-logo.svg',
+    alt: 'Future Of Libraries Is Open',
+  },
+  favicon: {
+    src: '@folio/stripes-core/default-assets/favicon.ico',
+  },
+};