+++ src/components/Front.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Link from 'react-router-dom/Link';
+import Headline from '@folio/stripes-components/lib/Headline';
 import { FormattedMessage } from 'react-intl';
 
 // TODO: release version of @folio/stripes-components has a broken index.js
@@ -9,14 +9,14 @@ import { FormattedMessage } from 'react-intl';
 // import { Pluggable } from '@folio/stripes-components';
 import Pluggable from '@folio/stripes-components/lib/Pluggable';
 
+import css from './Front.css';
 import AddContext from '../AddContext';
 
 const Front = ({ stripes }) => (
   <AddContext context={{ stripes }}>
     <Pluggable type="frontpage">
-      <div>
-        <h3><FormattedMessage id="stripes-core.front.welcome" /></h3>
-        <p><Link to="/about"><FormattedMessage id="stripes-core.front.about" /></Link></p>
+      <div className={css.frontWrap}>
+        <Headline faded tag="h1" margin="none" className={css.frontTitle}><FormattedMessage id="stripes-core.front.welcome" /></Headline>
       </div>
     </Pluggable>
   </AddContext>