+++ src/components/Login/LoginCtrl.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'; // eslint-disable-line
 import PropTypes from 'prop-types';
 import { connect as reduxConnect } from 'react-redux'; // eslint-disable-line
+import { SubmissionError } from 'redux-form';
 
 import { requestLogin, requestSSOLogin } from '../../loginServices';
 import Login from './Login';
@@ -8,12 +9,10 @@ import Login from './Login';
 class LoginCtrl extends Component {
   static contextTypes = {
     store: PropTypes.object,
-    router: PropTypes.object,
-    authFail: PropTypes.bool,
   }
 
   static propTypes = {
-    authFail: PropTypes.bool,
+    authFailure: PropTypes.string,
     ssoEnabled: PropTypes.bool,
     autoLogin: PropTypes.shape({
       username: PropTypes.string.isRequired,
@@ -24,12 +23,10 @@ class LoginCtrl extends Component {
   constructor(props, context) {
     super(props);
     this.store = context.store;
-    this.router = context.router;
     this.handleSubmit = this.handleSubmit.bind(this);
     this.sys = require('stripes-config'); // eslint-disable-line
     this.okapiUrl = this.sys.okapi.url;
     this.tenant = this.sys.okapi.tenant;
-    this.initialValues = { username: '', password: '' };
     this.handleSSOLogin = this.handleSSOLogin.bind(this);
     if (props.autoLogin && props.autoLogin.username) {
       this.handleSubmit(props.autoLogin);
@@ -37,10 +34,11 @@ class LoginCtrl extends Component {
   }
 
   handleSubmit(data) {
-    requestLogin(this.okapiUrl, this.store, this.tenant, data).then((response) => {
+    return requestLogin(this.okapiUrl, this.store, this.tenant, data).then((response) => {
       if (response.status >= 400) {
-        // eslint-disable-next-line no-param-reassign
-        this.initialValues.username = data.username;
+        // On login failure: Throw submission error
+        // which triggers redux-form submitFailed prop
+        throw new SubmissionError();
       }
     });
   }
@@ -50,13 +48,11 @@ class LoginCtrl extends Component {
   }
 
   render() {
-    const { authFail, ssoEnabled } = this.props;
-
+    const { authFailure, ssoEnabled } = this.props;
     return (
       <Login
         onSubmit={this.handleSubmit}
-        authFail={authFail}
-        initialValues={this.initialValues}
+        authError={authFailure}
         handleSSOLogin={this.handleSSOLogin}
         ssoActive={ssoEnabled}
       />
@@ -66,7 +62,7 @@ class LoginCtrl extends Component {
 
 function mapStateToProps(state) {
   return {
-    authFail: state.okapi.authFailure,
+    authFailure: state.okapi.authFailure,
     ssoEnabled: state.okapi.ssoEnabled,
   };
 }